RWACS_GUI/requirements.txt
glob2==0.7
matplotlib==3.6.2
numpy==1.24.1
pyserial==3.5
PySimpleGUI==4.60.4